+++ discord/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/discord/message/general/Resume.kt
@@ -8,10 +8,12 @@ open class Resume
     @Expose
     @SerializedName("token")
     var token: String = ""
+
     @Expose
     @SerializedName("session_id")
     var session: String = ""
+
     @Expose
     @SerializedName("seq")
-    var sequenceNumber: Int = 0
+    var sequenceNumber: Long? = null
 }